feat(rpc): harden JSON-RPC and split endpoints; add config; fix build
- New `/rpc/submit` & `/rpc/query`; enforce POST, JSON CT, body limit, timeout, concurrency
- Per-IP token-bucket rate limits (+trusted proxy CIDRs), mempool cap, HTTP/1.1-only
- Add `RpcHardeningConfig` and wire into `OrchestrationConfig` (sane defaults)
- Forge: use new endpoints; simplify submit logic
- Deps: +tower, tower-http, ipnetwork, dashmap; tidy cfg-if
- Fix: homomorphic `position` usize→u64; derive SCALE Encode/Decode for dummy types

+++ crates/homomorphic/src/computation/mod.rs
@@ -92,7 +92,7 @@ impl<CS: HomomorphicCommitmentScheme + Clone> HomomorphicComputation<CS> {
         &self,
         operation: &CommitmentOperation,
         result: &CS::Commitment,
-        position: usize,
+        position: u64, // FIX: Changed from usize to u64
     ) -> HomomorphicResult<HomomorphicProof<CS>> {
         self.create_proof_with_selector(operation, result, &Selector::Position(position))
     }
